requirements.txt
langchain
langchain-groq
playwright
beautifulsoup4
httpx
pandas
pydantic
python-dotenv
sqlalchemy
dnspython
